File Upload Component Fixes (CodeQL)
