+++ apps/web/app/components/UploadingItem.tsx
@@ -17,7 +17,7 @@ export function UploadingItem({
   return (
     <div className="flex w-full bg-background rounded-md pb-4.5 shadow-sm overflow-hidden">
       <div className="flex min-w-0 w-full flex-col mt-4.5 mx-4.5 gap-2">
-        <h3 className="font-semibold text-accent truncate">{name}</h3>
+        <h3 className="text-headline text-foreground truncate">{name}</h3>
         <p className="text-label truncate">
           Uploading {Math.min(100, Math.round(progress))}%
         </p>
